// api/src/modules/payment-information/constants.ts
// eslint-disable-next-line no-shadow
export enum BANKING_TYPE {
  WIRE = 'wire',
  PAYPAL = 'paypal',
  ISSUE = 'issue_check_us',
  DEPOSIT = 'deposit',
  PAYONNEER = 'payoneer',
  BITPAY = 'bitpay',
  PAXUM = 'paxum'
}

// eslint-disable-next-line no-shadow
export enum BANKING_SOURCE {
  PERFORMER = 'performer',
  STUDIO = 'studio'
}